Refactor network handling and command structure
- Removed the NetworkHandle struct and replaced its usage with NetworkManager across various modules.
- Consolidated publish replication command handling into a new protocol structure.
- Introduced a new operation response tracker service to manage operation states.
- Updated command data handling to improve error management and response tracking.
- Refactored the publish service to utilize the new network protocol requests.
- Cleaned up unused code and comments in services and commands.
- Adjusted context and controller implementations to align with the new network structure.

+++ src/services/operation_response_tracker.rs
@@ -12,8 +12,8 @@ type Result<T> = std::result::Result<T, NodeError>;
 pub struct OperationState<OperationRequestMessageData: Clone> {
     pub nodes_found: Vec<PeerId>,
     pub last_contacted_index: usize,
-    pub failed_number: usize,
-    pub completed_number: usize,
+    pub failed_number: u8,
+    pub completed_number: u8,
     pub request_message: RequestMessage<OperationRequestMessageData>,
 }
